Extracted Parts of NeuralNetworkTrainer into own files and classes

+++ JNF_NEAT/main.cpp
@@ -3,7 +3,7 @@
 #include <time.h>
 
 int main() {
-    srand(time(nullptr));
+    srand((size_t)time(nullptr));
 	std::cout << "Hello World" << std::endl;
 	return 0;
 }
